moved circle into a new folder
