+++ client/src/components/StatsGenerator.jsx
@@ -1,5 +1,5 @@
 import Questions from "./Profile/Circle&Questions/Questions";
-import Circle from "./Circle";
+import Circle from "./Profile/Circle&Questions/Circle";
 import { useState } from "react";
 import axios from "axios";
 import Profile from "./Profile/About/Profile";
